requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-filter==23.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
Markdown==3.4.4
packaging==23.1
Pillow==10.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==5.5.0
tzdata==2023.3
urllib3==2.0.4
